Closed 154-Add-permissions-to-work-with-book-comments

+++ src/BookCrossingBackEnd/Controllers/BookChildCommentsController.cs
@@ -1,7 +1,7 @@
 ﻿using Application.Dto.Comment.Book;
 using Application.Services.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace BookCrossingBackEnd.Controllers
@@ -11,16 +11,22 @@ namespace BookCrossingBackEnd.Controllers
     public class BookChildCommentsController : ControllerBase
     {
         private readonly IBookChildCommentService _childBookCommentService;
-
-        public BookChildCommentsController(IBookChildCommentService childBookCommentService)
+        private readonly IUserResolverService _userResolverService;
+        public BookChildCommentsController(IBookChildCommentService childBookCommentService, IUserResolverService userResolverService)
         {
             _childBookCommentService = childBookCommentService;
+            _userResolverService = userResolverService;
         }
        
         // PUT: api/BookChildCommants
         [HttpPut]
+        [Authorize]
         public async Task<ActionResult<int>> Put([FromBody] ChildUpdateDto updateDto)
         {
+            if (updateDto.CommentOwnerId != _userResolverService.GetUserId())
+            {
+                return Forbid();
+            }
             int number = await _childBookCommentService.Update(updateDto);
             if (number == 0)
             {
@@ -31,10 +37,11 @@ namespace BookCrossingBackEnd.Controllers
 
         // POST: api/BookChildCommants
         [HttpPost]
+        [Authorize]
         public async Task<ActionResult<int>> Post([FromBody] ChildInsertDto insertDto)
         {
+            insertDto.CommentOwnerId = _userResolverService.GetUserId();
             int number = await _childBookCommentService.Add(insertDto);
-
             if (number == 0)
             {
                 return NotFound();
@@ -44,10 +51,14 @@ namespace BookCrossingBackEnd.Controllers
 
         // DELETE: api/BookChildCommants
         [HttpDelete]
+        [Authorize]
         public async Task<ActionResult<int>> Delete([FromBody] ChildDeleteDto deleteDto)
         {
+            if (deleteDto.CommentOwnerId != _userResolverService.GetUserId() && !_userResolverService.IsUserAdmin())
+            {
+                return Forbid();
+            }
             int number = await _childBookCommentService.Remove(deleteDto);
-
             if (number == 0)
             {
                 return NotFound();
